Remove "Find normalized_message in Logstash" link
This is convenient, but also often forgotten to be removed when the
normalized_message contains PII, because it's twice in the task description
(once plainly, and once within this link).

The main link that's importnat here, I think, is the reqId one because
that's hard to make a fast query for by hand. the normalised_message
one is easy enough to copy to an open mediawiki logstash board.

Change-Id: I60ba8fc5511ebcc7d50674fd5f033a5a81436454